refactor all calls for action types to refere to the types.js file, add finance reducer to root reducer
