+++ src/reducers/chores_reducer.js
@@ -1,9 +1,9 @@
-import {FETCH_CHORES} from '../actions'
+import { FETCH_CHORES } from '../actions/types'
 
 const INITIAL_STATE = []
 //research this
 
-export default function (state = INITIAL_STATE, action) {
+export default function(state = INITIAL_STATE, action) {
   switch(action.type) {
     case FETCH_CHORES:
       return [...state, ...action.payload.data]
